requirements.txt
torch
torchvision
ultralytics
opencv-python
numpy
pandas
matplotlib
streamlit
flask
pyyaml
